3743 - select division - code review updates
