+++ app/controllers/admin/admin_controller.rb
@@ -1,4 +1,7 @@
 class Admin::AdminController < ApplicationController
+  include DivisionSelectable
+  helper_method :selected_division_id
+
   before_action :authenticate_user!
   after_action :verify_authorized
 end
