+++ app/helpers/admin_helper.rb
@@ -1,0 +1,5 @@
+module AdminHelper
+  def division_select_options
+    [['All', nil]].concat(current_user.accessible_divisions.map{ |d| [d.name, d.id] })
+  end
+end
